# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 교수업적평가시스템 공통기능 API
  version: 0.1.0
servers:
- url: /api
security:
- SessionCookieAuth: []
...
paths:
...
  /api/auth/login:
    post:
      operationId: login
      summary: 로그인
      x-related-requirements:
      - REQ-038
      - REQ-057
      - REQ-058
      - REQ-059
      x-business-rules:
      - 상세코드 계층·연계 속성 관리 요청은 R09 세션과 서버 권한 검증을 통과해야 한다
      - 입력 필드는 Requirement Registry의 REQ-038 source-backed 항목만 허용한다
      x-side-effects:
      - 로그인 결과를 관련 로컬 DB 테이블에 저장하고 후속 조회에 반영한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 change_history에 기록할 수 있는 구조로 남긴다
      x-required-tests:
      - '{''happy'': ''유효한 R09 요청 -> 관련 테이블 변경분이 후속 조회에서 확인된다''}'
      - '{''auth'': ''R01 세션 요청 -> 403 ApiError와 데이터 무변경''}'
      - '{''validation'': ''필수 필드 누락 -> 해당 필드 오류가 ApiError.meta에 표시된다''}'
      - '{''business'': ''상세코드 계층·연계 속성 관리의 source-backed 제약 위반 -> 기존 상태가 유지된다''}'
      - '{''side-effect'': ''성공 요청 -> change_history 또는 관련 이력 구조에 처리자와 처리일시가 남는다''}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/CurrentUser'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
  /api/auth/logout:
    post:
      operationId: logout
      summary: 로그아웃
      x-related-requirements:
      - REQ-039
      x-roles:
      - R09
      x-business-rules:
      - 로그인 요청은 R09 세션과 서버 권한 검증을 통과해야 한다
      - 입력 필드는 Requirement Registry의 REQ-039 source-backed 항목만 허용한다
      x-side-effects:
      - 로그아웃 결과를 관련 로컬 DB 테이블에 저장하고 후속 조회에 반영한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 change_history에 기록할 수 있는 구조로 남긴다
      x-required-tests:
      - '{''happy'': ''유효한 R09 요청 -> 관련 테이블 변경분이 후속 조회에서 확인된다''}'
      - '{''auth'': ''R01 세션 요청 -> 403 ApiError와 데이터 무변경''}'
      - '{''validation'': ''필수 필드 누락 -> 해당 필드 오류가 ApiError.meta에 표시된다''}'
      - '{''business'': ''로그인의 source-backed 제약 위반 -> 기존 상태가 유지된다''}'
      - '{''side-effect'': ''성공 요청 -> change_history 또는 관련 이력 구조에 처리자와 처리일시가 남는다''}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/GenericObject'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
  /api/auth/me:
    get:
      operationId: getCurrentUser
      summary: 현재 사용자
      x-related-requirements:
      - REQ-040
      x-roles:
      - R09
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/CurrentUser'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
  /api/navigation/menus:
    get:
      operationId: listNavigationMenus
      summary: 사용자 메뉴 노출
      x-related-requirements:
      - REQ-022
      - REQ-023
      x-roles:
      - R09
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/MenuList'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      - name: filter
        in: query
        required: false
        schema:
          type: string
  /api/users:
    get:
      operationId: listUsers
      summary: 사용자 검색
      x-related-requirements:
      - REQ-001
      - REQ-002
      x-roles:
      - R09
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/SizeParam'
      - $ref: '#/components/parameters/SortParam'
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      - name: filter
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/UserList'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
  /api/users/{userId}/usage:
    patch:
      operationId: updateUserUsage
      summary: 사용자 사용여부 변경
      x-related-requirements:
      - REQ-003
      x-roles:
      - R09
      x-business-rules:
      - 사용자 사용여부 변경 요청은 R09 세션과 서버 권한 검증을 통과해야 한다
      - 입력 필드는 Requirement Registry의 REQ-003 source-backed 항목만 허용한다
      x-side-effects:
      - 사용자 사용여부 변경 결과를 관련 로컬 DB 테이블에 저장하고 후속 조회에 반영한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 change_history에 기록할 수 있는 구조로 남긴다
      x-state-transitions:
      - 사용자 사용여부 변경에 의해 대상 상태 또는 사용여부가 요청 값으로 전이된다
      x-required-tests:
      - '{''happy'': ''유효한 R09 요청 -> 관련 테이블 변경분이 후속 조회에서 확인된다''}'
      - '{''auth'': ''R01 세션 요청 -> 403 ApiError와 데이터 무변경''}'
      - '{''validation'': ''필수 필드 누락 -> 해당 필드 오류가 ApiError.meta에 표시된다''}'
      - '{''business'': ''사용자 사용여부 변경의 source-backed 제약 위반 -> 기존 상태가 유지된다''}'
      - '{''side-effect'': ''성공 요청 -> change_history 또는 관련 이력 구조에 처리자와 처리일시가 남는다''}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUsageRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/User'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
  /api/users/{userId}/business-roles:
    patch:
      operationId: updateUserBusinessRoles
      summary: 사용자 업무 역할 변경
      x-related-requirements:
      - REQ-004
      x-roles:
      - R09
      x-business-rules:
      - 사용자 업무 역할 변경 요청은 R09 세션과 서버 권한 검증을 통과해야 한다
      - 입력 필드는 Requirement Registry의 REQ-004 source-backed 항목만 허용한다
      x-side-effects:
      - 사용자 업무 역할 변경 결과를 관련 로컬 DB 테이블에 저장하고 후속 조회에 반영한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 change_history에 기록할 수 있는 구조로 남긴다
      x-required-tests:
      - '{''happy'': ''유효한 R09 요청 -> 관련 테이블 변경분이 후속 조회에서 확인된다''}'
      - '{''auth'': ''R01 세션 요청 -> 403 ApiError와 데이터 무변경''}'
      - '{''validation'': ''필수 필드 누락 -> 해당 필드 오류가 ApiError.meta에 표시된다''}'
      - '{''business'': ''사용자 업무 역할 변경의 source-backed 제약 위반 -> 기존 상태가 유지된다''}'
      - '{''side-effect'': ''성공 요청 -> change_history 또는 관련 이력 구조에 처리자와 처리일시가 남는다''}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessRoleRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/User'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
  /api/organizations:
    get:
      operationId: listOrganizations
      summary: 조직 검색
      x-related-requirements:
      - REQ-005
      x-roles:
      - R09
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/SizeParam'
      - $ref: '#/components/parameters/SortParam'
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      - name: filter
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/OrganizationList'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
  /api/organizations/tree:
    get:
      operationId: getOrganizationTree
      summary: 조직 계층
      x-related-requirements:
      - REQ-006
      x-roles:
      - R09
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/OrganizationTree'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
  /api/organization-relations:
    post:
      operationId: createOrganizationRelation
      summary: 조직 관계 적용기간 저장
      x-related-requirements:
      - REQ-007
      - REQ-008
      x-roles:
      - R09
      x-business-rules:
      - 조직 관계 적용기간 변경 요청은 R09 세션과 서버 권한 검증을 통과해야 한다
      - 입력 필드는 Requirement Registry의 REQ-007 source-backed 항목만 허용한다
      x-side-effects:
      - 조직 관계 적용기간 저장 결과를 관련 로컬 DB 테이블에 저장하고 후속 조회에 반영한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 change_history에 기록할 수 있는 구조로 남긴다
      x-required-tests:
      - '{''happy'': ''유효한 R09 요청 -> 관련 테이블 변경분이 후속 조회에서 확인된다''}'
      - '{''auth'': ''R01 세션 요청 -> 403 ApiError와 데이터 무변경''}'
      - '{''validation'': ''필수 필드 누락 -> 해당 필드 오류가 ApiError.meta에 표시된다''}'
      - '{''business'': ''조직 관계 적용기간 변경의 source-backed 제약 위반 -> 기존 상태가 유지된다''}'
      - '{''side-effect'': ''성공 요청 -> change_history 또는 관련 이력 구조에 처리자와 처리일시가 남는다''}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRelationRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Organization'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: 요청 오류
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
...
  /api/batch-reprocess-runs/{reprocessExecutionId}:
    get:
      operationId: getBatchReprocessRun
      summary: 재처리 결과 조회
      security:
      - SessionCookieAuth: []
      x-roles:
      - R09
      x-related-requirements:
      - REQ-261
      - REQ-267
      - REQ-275
      - REQ-279
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/SizeParam'
      - name: filter
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: KNUESESSION
...
  schemas:
    ApiResponse:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - true
        meta:
          type: object
          additionalProperties: true
        data: {}
    ApiError:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - false
        meta:
          type: object
          additionalProperties: true
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            fieldErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    GenericObject:
      type: object
      additionalProperties: true
    GenericRequest:
      type: object
      additionalProperties: true
...
    LoginRequest:
      type: object
      required:
      - loginId
      - password
      properties:
        loginId:
          type: string
        password:
          type: string
    CurrentUser:
      type: object
      properties:
        userId:
          type: string
        loginId:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleCode'
...
    MenuList:
      type: array
      items:
        $ref: '#/components/schemas/Menu'
...